==> 1-getting-data.ipynb <==
# Fetching data from API

First, let's inspect the data and POST request format.

import numpy as np 
import pandas as pd 
import requests

data = pd.read_csv("../input/dialect-dataset/dialect_dataset.csv")

data.head()

data.groupby('dialect').count()

data.info()

The API needs the id to be a string.

data =data.astype({"id": str})

Used yield not for a particular reason, I just wanted to practice more on using it.

Max_length = 458000
url = "https://recruitment.aimtechnologies.co/ai-tasks"
def getting_1000_texts():
    count = 0
    

    while True:
        if count == 458000:
            break
        id_list = data['id'][count:count+1000].to_list()
        dialect_list = data['dialect'][count:count+1000].to_list()
        response = requests.post(url, json =id_list)
        text_list = list(response.json().values())
        
        yield id_list, dialect_list, text_list
        count += 1000
        

ID_list = []
dialect_lis = []
text_lis = []

for ID, dialect, text in getting_1000_texts():
    ID_list.extend(ID)
    dialect_lis.extend(dialect)
    text_lis.extend(text)

### Getting the last 197 tweet

# getting the last 197 tweet
id_list = data['id'][Max_length:].to_list()
dialect_list = data['dialect'][Max_length:].to_list()
response = requests.post(url, json =id_list)
text_list = list(response.json().values())
ID_list.extend(id_list)
dialect_lis.extend(dialect_list)
text_lis.extend(text_list)
len(ID_list)

dict_lists = {'id': ID_list, 'dialect': dialect_lis, 'text': text_lis} 
    
df_text = pd.DataFrame(dict_lists)

df_text.head()

df_text.to_csv("dialect_data.csv",index=False)
