==> 2-preprocessing.ipynb <==
# Arabic Language Pre-Processing

In this notebook I'll clean the data and split it to train,val,test for using with different models.

import pandas as pd 
import numpy as np
from sklearn.model_selection import train_test_split
import re
import string
import emoji

data = pd.read_csv("../input/arabic-dialect/dialect_data.csv",lineterminator='\n')
data.head()

data.info()

## Helper Functions and variables for cleaning


# I don't need numbers and some tweets contained arabic numbers so I added them as punctuations.
arabic_punctuations = '''`÷×؛<>()*&^%][،/:"؟.,'{}~¦+|!”…“–•۱۲۳٤٥٦٧۸۹٠'''
translator = str.maketrans('', '', arabic_punctuations)

def give_emoji_free_text(text):
    # this function is better than encoding emojis pattern because of optimization.
    return emoji.get_emoji_regexp().sub(r'', text)

def normalizeArabic(text):
    text = text.strip()
    text = re.sub("[إأٱآا]", "ا", text)
    text = re.sub("ى", "ي", text)
    text = re.sub("ؤ", "ء", text)
    text = re.sub("ئ", "ء", text)
    text = re.sub("ة", "ه", text)
    text = re.sub("گ", "ك", text)
    text = ''.join(ch for ch in text)
    return text


def pre_processing(text):

    # removing longitaion
    text = re.sub(r'(.)\1+', r"\1\1", text)
    # remove hashtags
    text = re.sub(r'#', '', text)
    # removing user name
    text = re.sub('@[^\s]+', ' ', text)
    #Convert www.* or https?://* to " "
    text = re.sub('((www\.[^\s]+)|(https?://[^\s]+))',' ',text)
    
    # removing \n because I found it in some tweets while insepecting the data
    text = re.sub(r"\n"," ",text)
    # removing english characters and numbers
    text = re.sub(r'[A-Za-z0-9]', '', text)
    # removing puntication at last because of # and @
    text = text.translate(translator)

    # this one to remove underscore from text because it was affecting hashtags when it was in puntications
    text = re.sub(r'_',' ',text)

    # removing emoji and tashikel
    text = give_emoji_free_text(text)
    noise = re.compile(""" ّ    | # Tashdid
                             َ    | # Fatha
                             ً    | # Tanwin Fath
                             ُ    | # Damma
                             ٌ    | # Tanwin Damm
                             ِ    | # Kasra
                             ٍ    | # Tanwin Kasr
                             ْ     # Sukun
                         """, re.VERBOSE)
    text = re.sub(noise, '', text)

    
    # normalizing text
    text = normalizeArabic(text)
    
    # removing space more than 1
    text = re.sub(r'\s{2,}', ' ', text)
    
    return text

text = '• ۱۲۳٠ ﻋﻤﺮﻱ ﻣﺎ ﻓﻜﺮﺕ ﺃﺑﻴﻊ ﺍﻷﻳﺎﻡ ﺍﻟﺤﻠﻮﺓ ﺍﻟﻠﻲ ﻋﺸﺘﻬﺎ ﻣﻊ ﺃﻱ ﺷﺨﺺ . ﻣﻤﻜﻦ ﺃﻧﺴﺤﺐ، ﺃﺗﺮﺍﺟﻊ، ﺃﺯﻋﻞ ﻭﺃﻛﺮﻩ ﺑﺸﻜﻞ ﻣﺆﻗﺖ، ﻟﻜﻦ ﻋﻤﺮﻱ ﻣﺎ ﻓﻜﺮﺕ ﺃﺑﻴﻊ . ﺍﻟﻤﺴﺄﻟﺔ ﻣﺴﺄﻟﺔ ﻣﺒﺪﺃ 😸🥀'
pre_processing(text)

## Applying to all text

data['clean_text'] = data['text'].apply(lambda x: pre_processing(str(x))) 

## Inspecting data more

# this cell to make sure that there is no length == 0 in any text!

data['length'] = data['clean_text'].apply(lambda x: len(x)) 

data[data['length']<13]

### tweets with length less than 12 seems unimportant so I'll drop them.

final_data = data[data['length']>11]

#final_data.info()
# dropping the text before cleaning column

final_data.drop("text", axis=1, inplace=True)
final_data.info()

### Spliting the data for train, val, test.

### First let's split train and (val + test) 60%-40% then val and test 20%-20%

train_df,test_val_df = train_test_split(final_data, test_size=0.2, random_state=42,stratify = final_data['dialect'])
test_df,val_df = train_test_split(test_val_df, test_size=0.5, random_state=42,stratify = test_val_df['dialect'])

print(len(train_df))
print(len(test_df))
print(len(val_df))

# save them in csv files for next steps

train_df.to_csv("train.csv",index=False)
test_df.to_csv("test.csv",index=False)
val_df.to_csv("validation.csv",index=False)

### Pre-Processing techniques used:

- Removing all numbers, emojis and tashkill because we don't need them in predicting the dialect.
- Normalize the text.
- Removing punctautions.
