==> 3b-dlmodel.ipynb <==
# Deep Learning Model for Arabic dialect classification

First, let's import libraries and get the data.

import keras
import pandas as pd
from keras.callbacks import TensorBoard
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from sklearn.preprocessing import LabelEncoder
from keras.models import Sequential
from keras import layers
from keras.callbacks import EarlyStopping
import numpy as np

train = pd.read_csv("../input/clean-dialect-text/train.csv",lineterminator='\n')
val = pd.read_csv("../input/clean-dialect-text/validation.csv",lineterminator='\n')

y = train['dialect']
X = train['clean_text']
y_val = val['dialect']
x_val = val['clean_text']

## 1- Convert input-data from text to tokens and output-data to labels

tokenizer = Tokenizer()
tokenizer.fit_on_texts(X)

X_train_tokens = tokenizer.texts_to_sequences(X)
X_val_tokens = tokenizer.texts_to_sequences(x_val)
vocab_size = len(tokenizer.word_index) + 1 

vocab_size

encoder = LabelEncoder()

y_train = encoder.fit_transform(y)
y_val = encoder.transform(y_val)

y_val[0:5]

## 2- Padding the Input

We need to make sure that input from both validation and input are on the same length.

For this I checked the longest sequence and adding more to it for all.

# max_length = 0

# for lis in X_train_tokens:
#     length = len(lis)
#     if length > max_length:
#         max_length = length
    
    
# print(max_length)

maxlen = 100

X_train = pad_sequences(X_train_tokens, padding='post', maxlen=maxlen)
X_val = pad_sequences(X_val_tokens, padding='post', maxlen=maxlen)

## 3- Making the model

- Embedding size: choosed by trying for different numbers between 100-300
- After trying different layers, it showed that BILSTM is slightly better.
- There is no f1_score in keras so I'll use tensorflow addons.

# import tensorflow_addons as tfa 

# f1 = tfa.metrics.F1Score( average='macro')

embedding_dim = 200

model = Sequential()
model.add(layers.Embedding(input_dim=vocab_size, 
                           output_dim=embedding_dim, 
                           input_length=maxlen))
model.add(layers.Bidirectional(keras.layers.LSTM(64)))
#model.add(layers.Dropout(0.1))
model.add(layers.Dense(64, activation='relu',kernel_initializer='he_uniform'))
model.add(layers.Dropout(0.5))
model.add(layers.Dense(32, activation='relu'))
model.add(layers.Dense(18,activation='softmax'))

model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy',
              metrics=['acc'])
model.summary()



## 4- Fitting the model

- epochs = 4, because more than that the model overfits.
- Batch size = 256 , also choosed by trial and error.

history = model.fit(X_train, y_train,
                    epochs=4,
                    verbose=True,
                    validation_data=(X_val, y_val),
                    batch_size=256)
                    #callbacks=[EarlyStopping(monitor='val_loss',min_delta=0.05)])


## Checking F1-score on test

test = pd.read_csv("../input/clean-dialect-text/test.csv",lineterminator='\n')
y_test = test['dialect']
x_test = test['clean_text']
X_test = tokenizer.texts_to_sequences(x_test)
Y_test = encoder.transform(y_test)
X_test = pad_sequences(X_test, padding='post', maxlen=maxlen)

output = model.predict(X_test)
out = output.argmax(axis = 1)[:,None]
# reshaping to enter the f1 function
Y_pred = out.reshape((out.shape[0],))
Y_true = np.array(Y_test)

from sklearn.metrics import f1_score

f1_score(Y_true, Y_pred, average='macro')

# print(np.argmax(output))
# list(encoder.inverse_transform([np.argmax(output)]))

## 5- Saving the model, tokenizer and the encoder

The model:

model_json = model.to_json()
with open("model.json", "w") as json_file:
    json_file.write(model_json)
# serialize weights to HDF5
model.save_weights("model.h5")
print("Saved model to disk")

Tokenizer:

import pickle

# saving
with open('tokenizer.pickle', 'wb') as handle:
    pickle.dump(tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)

Encoder:

with open('encoder.pkl', 'wb') as handle:
    pickle.dump(encoder,handle)

## Conclusions:

- Because Arabic dialects are similar, it'll be better if we used less classes.
- I believe that ML model is better.
